ai-flow/requirements.txt
# AI Flow - Dependencies
# LangChain & LangGraph for agent orchestration
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0

# Google Generative AI (Free tier)
google-generativeai>=0.8.0

# Ollama for local LLM (optional)
ollama>=0.3.0

# Vector store
pgvector>=0.3.0
psycopg2-binary>=2.9.0

# Utilities
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0

# File operations
aiofiles>=24.0.0
gitpython>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
